Add notes about MetricExporter naming (#2286)
If the implementors choose to implement pull metric exporters modeled as a metric reader, the only variant for metric exporter interface is push metric exporter. While the pull metric exporter is modeled as a "metric reader", the better naming for the genric "metric exporter" interface may be "push metric exporter".

Just open this PR to see if there are similar naming confusions with other SDK implementors. If so, I'd believe the metric exporter and metric reader section need a re-structure in a follow-up PR since how the pull metric exporter is modeled is important to the naming of the metric exporter interface.

Related:
- https://github.com/open-telemetry/opentelemetry-js/pull/2707
- https://github.com/open-telemetry/opentelemetry-js/issues/2589#issuecomment-1007723767